+++ influxdb3/point_value_iterator_test.go
@@ -58,6 +58,7 @@ func TestPointValueIterator(t *testing.T) {
 		{Name: "f15", Type: arrow.FixedWidthTypes.MonthInterval},
 		{Name: "f16", Type: arrow.FixedWidthTypes.DayTimeInterval},
 		{Name: "f17", Type: arrow.FixedWidthTypes.Duration_s},
+		{Name: "f18", Type: arrow.Null},
 	}, nil)
 
 	var buf bytes.Buffer
@@ -87,6 +88,7 @@ func TestPointValueIterator(t *testing.T) {
 	rb.Field(15).(*array.MonthIntervalBuilder).Append(arrow.MonthInterval(int32(15)))
 	rb.Field(16).(*array.DayTimeIntervalBuilder).AppendNull()
 	rb.Field(17).(*array.DurationBuilder).Append(arrow.Duration(int64(17)))
+	rb.Field(18).(*array.NullBuilder).AppendNull()
 
 	rec = rb.NewRecord()
 	_ = writer.Write(rec)
@@ -109,6 +111,7 @@ func TestPointValueIterator(t *testing.T) {
 	rb.Field(15).(*array.MonthIntervalBuilder).Append(arrow.MonthInterval(int32(15)))
 	rb.Field(16).(*array.DayTimeIntervalBuilder).AppendNull()
 	rb.Field(17).(*array.DurationBuilder).Append(arrow.Duration(int64(17)))
+	rb.Field(18).(*array.NullBuilder).AppendNull()
 
 	rec = rb.NewRecord()
 	_ = writer.Write(rec)
@@ -142,6 +145,7 @@ func TestPointValueIterator(t *testing.T) {
 	var resultSet15 []any
 	var resultSet16 []any
 	var resultSet17 []any
+	var resultSet18 []any
 
 	for {
 		pointValues, err := it.Next()
@@ -169,6 +173,7 @@ func TestPointValueIterator(t *testing.T) {
 		resultSet15 = append(resultSet15, pointValues.GetField("f15"))
 		resultSet16 = append(resultSet16, pointValues.GetField("f16"))
 		resultSet17 = append(resultSet17, pointValues.GetField("f17"))
+		resultSet18 = append(resultSet18, pointValues.GetField("f18"))
 	}
 
 	assert.True(t, slices.Equal([]int64{0, 0}, resultSet0))
@@ -223,6 +228,9 @@ func TestPointValueIterator(t *testing.T) {
 
 	assert.Equal(t, resultSet17[0], arrow.Duration(int64(17)))
 	assert.Equal(t, resultSet17[1], arrow.Duration(int64(17)))
+
+	assert.Nil(t, resultSet18[0])
+	assert.Nil(t, resultSet18[1])
 
 	pointValues, err := it.Next()
 	assert.Equal(t, 2, it.Index())